# tpcc: modify join query so it's supported by opt