@nvanbenschoten 2018-07-18
Yes, this is a rare transaction so I don't expect it to alter perf. Still, I'd like to see a comparison between the old and the new query before making the change. You can run with `--wait=false --mix=stockLevel=1` to isolate it.