@RaduBerinde 2018-07-19
I will run that. But I assume I need to fist run the default "mix" for a while to populate the database with data? Or is there some already populated backup I can restore?